refactored order-complete folder
