+++ packages/gafl-webapp-service/src/routes/routes.js
@@ -20,7 +20,7 @@ import contact from '../pages/contact/contact/route.js'
 import newsletter from '../pages/contact/newsletter/route.js'
 
 import termsAndConditions from '../pages/terms-and-conditions/route.js'
-import orderComplete from '../pages/order-complete/route.js'
+import orderComplete from '../pages/order-complete/order-complete/route.js'
 import licenceInformation from '../pages/order-complete/licence-information/route.js'
 import paymentCancelled from '../pages/payment/cancelled/route.js'
 import paymentFailed from '../pages/payment/failed/route.js'
